+++ packages/cloud-cognitive/src/components/AddSelect/AddSelect.js
@@ -20,7 +20,6 @@ import {
 import { Tearsheet, TearsheetNarrow } from '../../components/Tearsheet';
 import { NoDataEmptyState } from '../../components/EmptyStates/NoDataEmptyState';
 import { pkg } from '../../settings';
-
 const componentName = 'AddSelect';
 
 export let AddSelect = forwardRef(
@@ -33,9 +32,12 @@ export let AddSelect = forwardRef(
       inputPlaceholder,
       items,
       itemsLabel,
-      multiSelect,
+      multi,
+      noResultsDescription,
+      noResultsTitle,
       noSelectionDescription,
       noSelectionTitle,
+      onSearchFilter,
       open,
       title,
       ...rest
@@ -43,9 +45,6 @@ export let AddSelect = forwardRef(
     ref
   ) => {
     const blockClass = `${pkg.prefix}--add-select`;
-
-    const [selected] = useState(0);
-
     const commonTearsheetProps = {
       open,
       title,
@@ -54,6 +53,31 @@ export let AddSelect = forwardRef(
       closeIconDescription: 'temp description',
     };
 
+    // hooks
+    const [selected] = useState(0);
+    const [searchTerm, setSearchTerm] = useState('');
+
+    // handlers
+    const onSearchHandler = (e) => {
+      const { value } = e.target;
+      setSearchTerm(value);
+      if (onSearchFilter) {
+        onSearchFilter(value);
+      }
+    };
+
+    // data manipulation
+    const getFilteredItems = () => {
+      // if the user uses their own filter then they provide the filtered items
+      if (onSearchFilter) {
+        return items;
+      }
+      // by default, just filter results by their label from a single search term
+      return items.filter((i) => i.label.includes(searchTerm));
+    };
+    const filteredResults = getFilteredItems();
+
+    // sidebar
     const influencer = (
       <div className={`${blockClass}__influencer`}>
         <div className={`${blockClass}__influencer-header`}>
@@ -76,6 +100,7 @@ export let AddSelect = forwardRef(
       </div>
     );
 
+    // main content
     const body = (
       <>
         <div className={`${blockClass}__header`}>
@@ -83,37 +108,48 @@ export let AddSelect = forwardRef(
             id="temp-id"
             labelText="temp label"
             placeholder={inputPlaceholder}
+            value={searchTerm}
+            onChange={onSearchHandler}
           />
           <div className={`${blockClass}__items-label-container`}>
             <p className={`${blockClass}__items-label`}>{itemsLabel}</p>
             <Tag type="gray" size="sm">
-              {items.length}
+              {filteredResults.length}
             </Tag>
           </div>
         </div>
-        <StructuredListWrapper
-          selection
-          className={`${blockClass}__selections`}
-        >
-          <StructuredListBody>
-            {items.map((item) => (
-              <StructuredListRow key={item.id}>
-                <StructuredListCell>
-                  <p>{item.label}</p>
-                </StructuredListCell>
-              </StructuredListRow>
-            ))}
-          </StructuredListBody>
-        </StructuredListWrapper>
+        {filteredResults.length > 0 ? (
+          <StructuredListWrapper
+            selection
+            className={`${blockClass}__selections`}
+          >
+            <StructuredListBody>
+              {filteredResults.map((item) => (
+                <StructuredListRow key={item.id}>
+                  <StructuredListCell>
+                    <p>{item.label}</p>
+                  </StructuredListCell>
+                </StructuredListRow>
+              ))}
+            </StructuredListBody>
+          </StructuredListWrapper>
+        ) : (
+          <div className={`${blockClass}__body`}>
+            <NoDataEmptyState
+              subtitle={noResultsDescription}
+              title={noResultsTitle}
+            />
+          </div>
+        )}
       </>
     );
 
     return (
       <div ref={ref} className={cx(className, blockClass)} {...rest}>
-        {multiSelect ? (
+        {multi ? (
           <Tearsheet
             {...commonTearsheetProps}
-            influencer={multiSelect && influencer}
+            influencer={multi && influencer}
             influencerPosition="right"
           >
             {body}
@@ -126,20 +162,25 @@ export let AddSelect = forwardRef(
   }
 );
 
-AddSelect = pkg.checkComponentEnabled(AddSelect, componentName);
-AddSelect.displayName = componentName;
-
 AddSelect.propTypes = {
   actions: PropTypes.array,
   className: PropTypes.string,
   description: PropTypes.string,
   influencerTitle: PropTypes.string,
   inputPlaceholder: PropTypes.string,
-  items: PropTypes.array,
+  items: PropTypes.arrayOf(
+    PropTypes.shape({
+      label: PropTypes.string,
+      id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+    })
+  ),
   itemsLabel: PropTypes.string,
-  multiSelect: PropTypes.bool,
+  multi: PropTypes.bool,
+  noResultsDescription: PropTypes.string,
+  noResultsTitle: PropTypes.string,
   noSelectionDescription: PropTypes.string,
   noSelectionTitle: PropTypes.string,
+  onSearchFilter: PropTypes.func,
   open: PropTypes.bool,
   title: PropTypes.string,
 };
@@ -147,3 +188,5 @@ AddSelect.propTypes = {
 AddSelect.defaultProps = {
   items: [],
 };
+
+AddSelect.displayName = componentName;