+++ packages/cloud-cognitive/src/components/MultiAddSelect/MultiAddSelect.test.js
@@ -1,0 +1,39 @@
+/**
+ * Copyright IBM Corp. 2021
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import { render } from '@testing-library/react';
+import React from 'react';
+import { MultiAddSelect } from './MultiAddSelect';
+
+const componentName = MultiAddSelect.name;
+const defaultProps = {
+  noSelectionTitle: 'No selection title',
+  noSelectionDescription: 'No selection description',
+  noResultsTitle: 'No results title',
+  noResultsDescription: 'Try again description',
+};
+
+describe(componentName, () => {
+  const { ResizeObserver } = window;
+
+  beforeEach(() => {
+    window.ResizeObserver = jest.fn().mockImplementation(() => ({
+      observe: jest.fn(),
+      unobserve: jest.fn(),
+      disconnect: jest.fn(),
+    }));
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
+    window.ResizeObserver = ResizeObserver;
+  });
+
+  it('renders', () => {
+    render(<MultiAddSelect {...defaultProps} />);
+  });
+});