+++ packages/cloud-cognitive/src/global/js/package-settings.js
@@ -50,7 +50,8 @@ const defaults = {
     UserProfileImage: true,
 
     // other public components not yet reviewed and released:
-    AddSelect: false,
+    MultiAddSelect: false,
+    SingleAddSelect: false,
     LoadingBar: false,
     ModifiedTabs: false,
     Toolbar: false,