+++ packages/core/story-structure.js
@@ -24,7 +24,7 @@ const s = [
         n: 'Usage examples',
         s: [],
       },
-      'o/Examples on CodeSandbox',
+      'o/Examples',
     ],
   },
   {
@@ -46,7 +46,7 @@ const s = [
       {
         n: 'Patterns',
         s: [
-          { n: 'AddSelect', s: ['c/AddSelect'] },
+          { n: 'Add Select', s: ['c/SingleAddSelect', 'c/MultiAddSelect'] },
           { n: 'About modal', s: ['c/AboutModal'] },
           {
             n: 'Create flows',